feat(PHO-4146): finalize code
